refactor: change file structure; use local data if http request of couchdb fails
