--- web-app/src/components/Chart/ScenarioDistress/data/data_chart.json
@@ -1,127 +1,0 @@
-{
-  "Mel": [
-    { "positive": 61.2, "distress": 51.6 },
-    { "positive": 67.5, "distress": 59.0 },
-    { "positive": 59.5, "distress": 49.2 },
-    { "positive": 57.0, "distress": 63.0 },
-    { "positive": 55.8, "distress": 53.6 },
-    { "positive": 70.0, "distress": 59.0 },
-    { "positive": 59.1, "distress": 47.6 },
-    { "positive": 66.0, "distress": 69.8 },
-    { "positive": 76.2, "distress": 66.8 },
-    { "positive": 60.2, "distress": 75.2 },
-    { "positive": 72.5, "distress": 55.2 },
-    { "positive": 70.9, "distress": 54.2 },
-    { "positive": 72.9, "distress": 62.5 },
-    { "positive": 53.4, "distress": 42.0 },
-    { "positive": 60.0, "distress": 50.0 },
-    { "positive": 47.2, "distress": 49.8 },
-    { "positive": 68.2, "distress": 49.2 },
-    { "positive": 75.0, "distress": 73.2 },
-    { "positive": 57.0, "distress": 47.8 },
-    { "positive": 67.6, "distress": 68.8 },
-    { "positive": 59.5, "distress": 50.6 },
-    { "positive": 75.0, "distress": 82.5 },
-    { "positive": 66.8, "distress": 57.2 },
-    { "positive": 76.5, "distress": 87.8 },
-    { "positive": 70.2, "distress": 72.8 },
-    { "positive": 74.0, "distress": 54.5 },
-    { "positive": 73.0, "distress": 59.8 },
-    { "positive": 79.9, "distress": 67.3 },
-    { "positive": 70.5, "distress": 67.8 },
-    { "positive": 60.0, "distress": 47.0 },
-    { "positive": 54.4, "distress": 46.2 },
-    { "positive": 62.0, "distress": 55.0 }
-  ],
-  "Syd": [
-    { "positive": 74.0, "distress": 65.6 },
-    { "positive": 75.3, "distress": 71.8 },
-    { "positive": 93.5, "distress": 80.7 },
-    { "positive": 86.5, "distress": 72.6 },
-    { "positive": 87.2, "distress": 78.8 },
-    { "positive": 81.5, "distress": 74.8 },
-    { "positive": 84.0, "distress": 86.4 },
-    { "positive": 84.5, "distress": 78.4 },
-    { "positive": 75.0, "distress": 62.0 },
-    { "positive": 84.0, "distress": 81.6 },
-    { "positive": 80.0, "distress": 76.6 },
-    { "positive": 77.8, "distress": 83.6 },
-    { "positive": 92.0, "distress": 90.0 },
-    { "positive": 76.0, "distress": 74.6 },
-    { "positive": 74.0, "distress": 71.0 },
-    { "positive": 84.0, "distress": 79.6 },
-    { "positive": 92.7, "distress": 93.8 },
-    { "positive": 71.5, "distress": 70.0 },
-    { "positive": 73.0, "distress": 72.4 },
-    { "positive": 76.0, "distress": 85.9 },
-    { "positive": 76.0, "distress": 78.8 },
-    { "positive": 80.5, "distress": 77.8 },
-    { "positive": 72.7, "distress": 66.2 },
-    { "positive": 76.0, "distress": 86.4 },
-    { "positive": 73.5, "distress": 81.8 },
-    { "positive": 78.0, "distress": 89.6 },
-    { "positive": 80.3, "distress": 82.8 },
-    { "positive": 80.3, "distress": 76.4 },
-    { "positive": 64.5, "distress": 63.2 },
-    { "positive": 73.0, "distress": 60.9 },
-    { "positive": 83.5, "distress": 74.8 },
-    { "positive": 75.5, "distress": 70.0 },
-    { "positive": 88.0, "distress": 72.4 },
-    { "positive": 89.2, "distress": 84.1 },
-    { "positive": 72.8, "distress": 69.1 },
-    { "positive": 70.0, "distress": 59.5 },
-    { "positive": 82.0, "distress": 67.2 },
-    { "positive": 70.0, "distress": 61.3 },
-    { "positive": 77.8, "distress": 68.6 },
-    { "positive": 84.2, "distress": 80.1 },
-    { "positive": 86.7, "distress": 87.8 },
-    { "positive": 71.4, "distress": 84.7 },
-    { "positive": 72.7, "distress": 73.4 },
-    { "positive": 75.3, "distress": 72.1 },
-    { "positive": 80.3, "distress": 82.6 },
-    { "positive": 82.9, "distress": 88.7 },
-    { "positive": 88.0, "distress": 84.1 },
-    { "positive": 77.2, "distress": 94.1 },
-    { "positive": 72.1, "distress": 74.9 },
-    { "positive": 67.0, "distress": 59.1 },
-    { "positive": 69.5, "distress": 75.6 },
-    { "positive": 74.0, "distress": 86.2 },
-    { "positive": 72.7, "distress": 75.3 },
-    { "positive": 82.2, "distress": 87.1 },
-    { "positive": 64.1, "distress": 55.2 },
-    { "positive": 63.0, "distress": 57.0 },
-    { "positive": 71.5, "distress": 61.4 }
-  ],
-  "Perth": [
-    { "positive": 84.2, "distress": 76.8 },
-    { "positive": 74.0, "distress": 86.8 },
-    { "positive": 74.0, "distress": 72.2 },
-    { "positive": 77.0, "distress": 71.6 },
-    { "positive": 86.0, "distress": 84.8 },
-    { "positive": 67.0, "distress": 68.2 },
-    { "positive": 71.8, "distress": 66.1 },
-    { "positive": 82.0, "distress": 72.0 },
-    { "positive": 67.0, "distress": 64.6 },
-    { "positive": 77.8, "distress": 74.8 },
-    { "positive": 64.5, "distress": 70.0 },
-    { "positive": 92.0, "distress": 101.6 },
-    { "positive": 75.5, "distress": 63.2 },
-    { "positive": 71.2, "distress": 79.1 },
-    { "positive": 81.6, "distress": 78.9 },
-    { "positive": 67.4, "distress": 67.7 },
-    { "positive": 81.1, "distress": 66.0 },
-    { "positive": 77.0, "distress": 68.2 },
-    { "positive": 74.5, "distress": 63.9 },
-    { "positive": 77.5, "distress": 72.0 },
-    { "positive": 70.5, "distress": 56.8 },
-    { "positive": 82.4, "distress": 74.5 },
-    { "positive": 97.1, "distress": 90.9 },
-    { "positive": 80.1, "distress": 93.0 },
-    { "positive": 75.5, "distress": 80.9 },
-    { "positive": 80.6, "distress": 72.7 },
-    { "positive": 84.4, "distress": 68.0 },
-    { "positive": 75.5, "distress": 70.9 },
-    { "positive": 80.6, "distress": 72.5 },
-    { "positive": 77.0, "distress": 72.5 }
-  ]
-}
