--- web-app/src/components/Chart/ScenarioSitting/data/data_bar.json
@@ -1,132 +1,0 @@
-[
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 },
-  { "LGAcode": 3052, "count": 510 },
-  { "LGAcode": 3054, "count": 590 },
-  { "LGAcode": 3067, "count": 190 },
-  { "LGAcode": 3022, "count": 234 },
-  { "LGAcode": 3023, "count": 111 },
-  { "LGAcode": 3099, "count": 256 },
-  { "LGAcode": 3021, "count": 435 },
-  { "LGAcode": 3024, "count": 345 },
-  { "LGAcode": 3025, "count": 234 },
-  { "LGAcode": 3026, "count": 445 }
-]
