--- web-app/src/components/Chart/ScenarioSitting/data/data_chart.json
@@ -1,266 +1,0 @@
-{
-  "name": "Melbourne",
-  "label": {},
-  "itemStyle": {},
-  "children": [
-    {
-      "name": "Parkville",
-      "value": 4116,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "Film",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Teng",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Fei",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Zhang",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Alick",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "Carlton",
-      "value": 41,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "Grace",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Sun",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Hao",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Ran",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Foolish",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "SouthBank",
-      "value": 4118,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "Shao",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Yun",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Peng",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Han",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Tim",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "Hawthorn",
-      "value": 413,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "wowowowowo",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "hahahahah",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Fei",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Teng",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "Boxhill",
-      "value": 412,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "wowowowowo",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "hahahahah",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Fei",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Teng",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "Docklands",
-      "value": 44,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "wowowowowo",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "hahahahah",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Fei",
-          "value": 43,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        },
-        {
-          "name": "Teng",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    },
-    {
-      "name": "Flemington",
-      "value": 419,
-      "label": {},
-      "itemStyle": {},
-      "collapsed": null,
-      "children": [
-        {
-          "name": "Film",
-          "value": 41,
-          "label": {},
-          "itemStyle": {},
-          "collapsed": null
-        }
-      ]
-    }
-  ]
-}
