--- web-app/src/components/Chart/index.js
@@ -1,8 +1,0 @@
-export { default as ScenarioDistressChart } from './ScenarioDistress/ScenarioDistressChart';
-export { default as ScenarioSafetyChart } from './ScenarioSafety/ScenarioSafetyChart';
-export { default as ScenarioSittingChart } from './ScenarioSitting/ScenarioSittingChart';
-// export { default as ScenarioSittingBar } from './ScenarioSitting/ScenarioSittingBar';
-export { default as ScenarioWeatherWindChart } from './ScenarioWeather/ScenarioWeatherWindChart';
-export { default as ScenarioWeatherRainChart } from './ScenarioWeather/ScenarioWeatherRainChart';
-export { default as ScenarioCityChart } from './ScenarioCity/ScenarioCityChart';
-export { default as ScenarioSittingScatter } from './ScenarioSitting/ScenarioSittingScatter';
