+++ web-app/src/components/City/City.js
@@ -1,5 +1,6 @@
 import React, { Component, Fragment } from 'react';
-import { ScenarioCityChart } from '../Chart';
+import ScenarioCityChart from './components/ScenarioCityChart';
+import dataChartJson from './data/data_chart.json';
 
 export default class City extends Component {
   render() {
@@ -11,7 +12,7 @@ export default class City extends Component {
             {`For different cities, there might be different cultures, so people might use different words. For example, those people who live near the equatorial may post more word such as “hot” etc. For those cities that are relatively close to the university, their high-frequency words may be "libraries". We grab the coordinates from the tweets we captured and then classify them into different cities. Count the tweets issued in each city to obtain high-frequency words for different cities.`}
           </div>
         </div>
-        <ScenarioCityChart />;
+        <ScenarioCityChart data={dataChartJson} />;
       </Fragment>
     );
   }
