+++ web-app/src/components/Distress/data/data_chart.json
@@ -1,0 +1,108 @@
+{
+  "melbourne": {
+    "27350": [1035, 330, 1250, "Yarra (C)", 11.0657480897472, 0.4],
+    "25060": [399, 110, 399, "Moonee Valley (C)", 10.5751840233999, 0.44],
+    "25900": [1236, 377, 1322, "Port Phillip (C)", 10.3502356455026, 0.42],
+    "21180": [420, 106, 343, "Brimbank (C)", 13.3245926332471, 0.48],
+    "21890": [538, 280, 606, "Darebin (C)", 12.93378039402, 0.38],
+    "23110": [228, 84, 316, "Hobsons Bay (C)", 11.8186786307422, 0.36],
+    "26350": [850, 211, 1031, "Stonnington (C)", 9.72536237313447, 0.41],
+    "23430": [389, 137, 442, "Kingston (C)", 11.292201605311, 0.4],
+    "21450": [149, 7, 44, "Cardinia (S)", 11.2914290577328, 0.74],
+    "21110": [646, 211, 830, "Boroondara (C)", 8.59651772566375, 0.38],
+    "22670": [222, 63, 209, "Greater Dandenong (C)", 14.3560068331146, 0.45],
+    "24410": [231, 96, 333, "Maroondah (C)", 10.5828072473045, 0.35],
+    "20910": [548, 127, 706, "Bayside (C)", 9.7092012221571, 0.4],
+    "27450": [408, 52, 439, "Yarra Ranges (S)", 10.361608621799, 0.45],
+    "22310": [781, 281, 1058, "Glen Eira (C)", 10.3078254048687, 0.37],
+    "21610": [208, 73, 381, "Casey (C)", 11.5319964384208, 0.31],
+    "27070": [226, 57, 475, "Whittlesea (C)", 12.0930405774231, 0.3],
+    "26980": [155, 54, 334, "Whitehorse (C)", 10.7889722530965, 0.29],
+    "23270": [547, 192, 920, "Hume (C)", 12.87946532008, 0.33],
+    "25250": [914, 278, 1136, "Moreland (C)", 11.5482803852727, 0.39],
+    "24210": [83, 51, 156, "Manningham (C)", 8.82647256451655, 0.29],
+    "24600": [9806, 5091, 16376, "Melbourne (C)", 12.4243670764612, 0.31],
+    "24650": [32, 14, 70, "Melton (S)", 10.4134886438296, 0.28],
+    "24970": [478, 470, 814, "Monash (C)", 10.5309212041225, 0.27],
+    "25710": [87, 7, 46, "Nillumbik (S)", 8.50105464270863, 0.62],
+    "23670": [267, 87, 400, "Knox (C)", 10.954544618462, 0.35],
+    "24330": [330, 85, 431, "Maribyrnong (C)", 12.0705116256603, 0.39],
+    "20660": [250, 72, 423, "Banyule (C)", 10.3718574452266, 0.34],
+    "29399": [512, 96, 518, "Unincorporated Vic", 11.6020874429726, 0.45],
+    "27260": [179, 60, 190, "Wyndham (C)", 10.50899899323, 0.42]
+  },
+  "sydney": {
+    "10750": [1051, 390, 1024, "Blacktown (C)", 10.750275449192, 0.43],
+    "18400": [360, 122, 566, "Wollondilly (A)", 9.64249554807054, 0.34],
+    "17200": [12149, 3142, 17767, "Sydney (C)", 10.278156288235, 0.37],
+    "14450": [158, 37, 168, "Kogarah (C)", 10.2715857739501, 0.44],
+    "18250": [808, 215, 1075, "Willoughby (C)", 7.99377721316972, 0.39],
+    "10150": [233, 52, 185, "Ashfield (A)", 10.6316292344196, 0.5],
+    "14000": [224, 95, 407, "Hornsby (A)", 8.59012445652474, 0.31],
+    "10200": [878, 604, 1120, "Auburn (C)", 11.9447945982131, 0.34],
+    "18000": [1082, 288, 1396, "Warringah (A)", 8.86238241677098, 0.39],
+    "14800": [639, 192, 834, "Leichhardt (A)", 8.78717064505151, 0.38],
+    "11100": [535, 190, 1120, "Botany Bay (C)", 11.6964026658011, 0.29],
+    "10350": [218, 59, 429, "Bankstown (C)", 12.2864465056634, 0.31],
+    "16650": [264, 46, 538, "Rockdale (C)", 10.2661046422642, 0.31],
+    "11550": [87, 20, 126, "Canterbury (C)", 11.066435284476, 0.37],
+    "14150": [118, 11, 117, "Hurstville (C)", 10.4071015156, 0.48],
+    "18050": [1389, 358, 1800, "Waverley (A)", 9.36952785686353, 0.39],
+    "12850": [124, 23, 159, "Fairfield (C)", 13.4613174466577, 0.41],
+    "17150": [801, 177, 1007, "Sutherland Shire (A)", 8.78370943526021, 0.4],
+    "11300": [61, 18, 141, "Burwood (A)", 10.7587925425434, 0.28],
+    "15950": [623, 256, 752, "North Sydney (A)", 8.48627241874895, 0.38],
+    "15200": [1401, 434, 1857, "Marrickville (A)", 9.4289753665951, 0.38],
+    "11520": [304, 40, 228, "Canada Bay (A)", 9.40443781233525, 0.53],
+    "16250": [755, 113, 775, "Parramatta (C)", 11.0164145338007, 0.46],
+    "17420": [387, 94, 288, "The Hills Shire (A)", 8.22253734900258, 0.5],
+    "16350": [406, 111, 669, "Penrith (C)", 10.0896201335758, 0.34],
+    "16550": [694, 164, 1019, "Randwick (C)", 10.2345916137126, 0.37],
+    "18500": [374, 55, 405, "Woollahra (A)", 9.26702027396741, 0.45],
+    "14100": [75, 24, 74, "Hunters Hill (A)", 9.42505994050728, 0.43],
+    "11500": [161, 23, 185, "Campbelltown (C)", 11.5036621720229, 0.44],
+    "13800": [96, 21, 117, "Hawkesbury (C)", 10.0073282952112, 0.41],
+    "17100": [40, 9, 41, "Strathfield (A)", 10.4614764893218, 0.44],
+    "14500": [137, 23, 156, "Ku-ring-gai (A)", 7.05125652224601, 0.43],
+    "16700": [365, 61, 831, "Ryde (C)", 9.32902908946225, 0.29],
+    "15150": [558, 80, 531, "Manly (A)", 9.13023964979741, 0.48],
+    "15350": [418, 54, 381, "Mosman (A)", 9.02880779779002, 0.49],
+    "11450": [185, 17, 203, "Camden (A)", 9.30158191619081, 0.46],
+    "16370": [627, 120, 540, "Pittwater (A)", 8.35573121601599, 0.49],
+    "13950": [29, 24, 74, "Holroyd (C)", 11.4084341694295, 0.23],
+    "14700": [109, 22, 85, "Lane Cove (A)", 7.58428945041326, 0.5],
+    "14900": [151, 11, 86, "Liverpool (C)", 11.5363559211486, 0.61]
+  },
+  "perth": {
+    "57840": [93, 14, 117, "South Perth (C)", 10.1801453521419, 0.42],
+    "57080": [1841, 495, 2315, "Perth (C)", 10.3083226283361, 0.4],
+    "50490": [112, 26, 215, "Belmont (C)", 11.6378957532254, 0.32],
+    "58510": [273, 28, 292, "Victoria Park (T)", 11.1019157454034, 0.46],
+    "52170": [50, 0, 25, "Cottesloe (T)", 9.46469593484165, 0.67],
+    "58570": [97, 22, 109, "Vincent (T)", 10.1846956629017, 0.43],
+    "54200": [206, 303, 176, "Kalamunda (S)", 8.87370753115408, 0.3],
+    "55320": [101, 7, 132, "Melville (C)", 9.84045942122714, 0.42],
+    "51330": [71, 38, 143, "Canning (C)", 10.8556177904622, 0.28],
+    "57910": [360, 85, 411, "Stirling (C)", 10.5663422475098, 0.42],
+    "50420": [73, 16, 71, "Bayswater (C)", 10.7787881828329, 0.46],
+    "51820": [324, 37, 486, "Cockburn (C)", 10.1861543773832, 0.38],
+    "53150": [16, 14, 80, "East Fremantle (T)", 10.1080748264504, 0.15],
+    "58760": [116, 41, 196, "Wanneroo (C)", 10.4147523122449, 0.33],
+    "51310": [149, 26, 193, "Cambridge (T)", 8.47559600751305, 0.4],
+    "58050": [65, 7, 94, "Swan (C)", 10.1437352648619, 0.39],
+    "56090": [13, 7, 18, "Mundaring (S)", 10.0755989588623, 0.34],
+    "53780": [21, 24, 37, "Gosnells (C)", 11.5578353051635, 0.26],
+    "57490": [83, 15, 105, "Rockingham (C)", 10.8438999026425, 0.41],
+    "50210": [69, 19, 76, "Armadale (C)", 11.2632947729896, 0.42],
+    "54830": [38, 0, 15, "Kwinana (T)", 11.1597141206203, 0.72],
+    "57980": [68, 3, 44, "Subiaco (C)", 9.69308480412188, 0.59],
+    "50350": [15, 7, 38, "Bassendean (T)", 11.752587931207, 0.25],
+    "51750": [109, 34, 104, "Claremont (T)", 9.08119628733736, 0.44],
+    "56580": [25, 11, 70, "Nedlands (C)", 7.63803974136651, 0.24],
+    "54170": [165, 16, 93, "Joondalup (C)", 9.58555524735361, 0.6],
+    "53430": [14, 11, 45, "Fremantle (C)", 10.9899425337873, 0.2],
+    "56930": [15, 7, 78, "Peppermint Grove (S)", 9.46469593484165, 0.15],
+    "55740": [14, 0, 11, "Mosman Park (T)", 9.46469593484165, 0.56],
+    "57700": [6, 0, 0, "Serpentine-Jarrahdale (S)", 8.36174680959785, 1.0]
+  }
+}
