+++ web-app/src/components/ScenarioList/ScenarioList.scss
@@ -45,11 +45,12 @@
   }
 
   .fifth-scenario {
-    width: 60%;
+    width: 65%;
     margin-left: 15%;
     flex-direction: row-reverse;
 
     .card-img-container {
+      width: 70%;
       margin-right: 1rem;
     }
   }
