+++ web-app/src/components/ScenarioList/images/scenario_sitting.jpg
