+++ web-app/src/components/Sitting/data/data_chart.json
@@ -1,0 +1,33 @@
+{
+  "27350": [1035, 330, 1250, "Yarra (C)", 11.0657480897472, 93.1, 73.8, 43.4, 0.4, 2615],
+  "25060": [399, 110, 399, "Moonee Valley (C)", 10.5751840233999, 97.2, 76.9, 29.5, 0.44, 908],
+  "25900": [1236, 377, 1322, "Port Phillip (C)", 10.3502356455026, 97.7, 78.6, 44.7, 0.42, 2935],
+  "21180": [420, 106, 343, "Brimbank (C)", 13.3245926332471, 89.6, 45.2, 24.9, 0.48, 869],
+  "21890": [538, 280, 606, "Darebin (C)", 12.93378039402, 95.9, 60.9, 40.2, 0.38, 1424],
+  "23110": [228, 84, 316, "Hobsons Bay (C)", 11.8186786307422, 95.6, 70.7, 42.3, 0.36, 628],
+  "26350": [850, 211, 1031, "Stonnington (C)", 9.72536237313447, 96.9, 80.4, 35.4, 0.41, 2092],
+  "23430": [389, 137, 442, "Kingston (C)", 11.292201605311, 0, 0, 0, 0.4, 968],
+  "21450": [149, 7, 44, "Cardinia (S)", 11.2914290577328, 0, 0, 0, 0.74, 200],
+  "21110": [646, 211, 830, "Boroondara (C)", 8.59651772566375, 99.5, 84.6, 45.9, 0.38, 1687],
+  "22670": [222, 63, 209, "Greater Dandenong (C)", 14.3560068331146, 0, 0, 0, 0.45, 494],
+  "24410": [231, 96, 333, "Maroondah (C)", 10.5828072473045, 97.6, 64.6, 33.8, 0.35, 660],
+  "20910": [548, 127, 706, "Bayside (C)", 9.7092012221571, 0, 0, 0, 0.4, 1381],
+  "27450": [408, 52, 439, "Yarra Ranges (S)", 10.361608621799, 99.0, 73.7, 27.2, 0.45, 899],
+  "22310": [781, 281, 1058, "Glen Eira (C)", 10.3078254048687, 98.3, 75.3, 40.7, 0.37, 2120],
+  "21610": [208, 73, 381, "Casey (C)", 11.5319964384208, 0, 0, 0, 0.31, 662],
+  "27070": [226, 57, 475, "Whittlesea (C)", 12.0930405774231, 93.6, 68.4, 29.0, 0.3, 758],
+  "None": [190, 67, 262, "", 0, 0, 0, 0, 0.37, 519],
+  "26980": [155, 54, 334, "Whitehorse (C)", 10.7889722530965, 98.2, 68.0, 41.5, 0.29, 543],
+  "23270": [547, 192, 920, "Hume (C)", 12.87946532008, 93.0, 56.1, 31.2, 0.33, 1659],
+  "25250": [914, 278, 1136, "Moreland (C)", 11.5482803852727, 96.3, 68.1, 38.1, 0.39, 2328],
+  "24210": [83, 51, 156, "Manningham (C)", 8.82647256451655, 97.4, 68.7, 31.6, 0.29, 290],
+  "24600": [9806, 5091, 16376, "Melbourne (C)", 12.4243670764612, 99.1, 78.0, 52.3, 0.31, 31273],
+  "24650": [32, 14, 70, "Melton (S)", 10.4134886438296, 95.9, 57.6, 29.0, 0.28, 116],
+  "24970": [478, 470, 814, "Monash (C)", 10.5309212041225, 98.1, 73.4, 33.1, 0.27, 1762],
+  "25710": [87, 7, 46, "Nillumbik (S)", 8.50105464270863, 98.6, 82.5, 32.2, 0.62, 140],
+  "23670": [267, 87, 400, "Knox (C)", 10.954544618462, 98.0, 62.4, 39.9, 0.35, 754],
+  "24330": [330, 85, 431, "Maribyrnong (C)", 12.0705116256603, 95.5, 64.1, 38.2, 0.39, 846],
+  "20660": [250, 72, 423, "Banyule (C)", 10.3718574452266, 98.2, 77.4, 34.6, 0.34, 745],
+  "29399": [512, 96, 518, "Unincorporated Vic", 11.6020874429726, 0, 0, 0, 0.45, 1126],
+  "27260": [179, 60, 190, "Wyndham (C)", 10.50899899323, 96.3, 54.2, 32.8, 0.42, 429]
+}
