+++ web-app/src/components/Weather/Weather.js
@@ -1,13 +1,15 @@
 import React, { Component } from 'react';
-import { ScenarioWeatherWindChart } from '../Chart';
-import { ScenarioWeatherRainChart } from '../Chart';
+import ScenarioWeatherWindChart from './components/ScenarioWeatherWindChart';
+import ScenarioWeatherRainChart from './components/ScenarioWeatherRainChart';
+import dataWindChartJson from './data/data_wind_chart.json';
+import dataRainChartJson from './data/data_rain_chart.json';
 
 export default class Weather extends Component {
   render() {
     return (
       <div>
-        <ScenarioWeatherWindChart />
-        <ScenarioWeatherRainChart />
+        <ScenarioWeatherWindChart data={dataWindChartJson} />
+        <ScenarioWeatherRainChart data={dataRainChartJson} />
       </div>
     );
   }
