+++ web-app/src/components/Weather/data/data_rain_chart.json
