+++ web-app/src/components/Weather/data/data_wind_chart.json
