+++ web-app/src/scenes/Scenario/Scenario.scss
@@ -23,14 +23,15 @@
     }
 
     .intro {
-      padding: 20px 20px 40px 20px;
+      padding: 40px;
       margin: 40px 0;
       background: #fff;
 
       .title {
+        margin-bottom: 20px;
         font-size: 2rem;
         font-weight: 700;
-        line-height: 2.2;
+        line-height: 1.2;
         text-align: center;
         text-decoration: uppercase;
       }
@@ -46,6 +47,10 @@
   .scenario {
     &-container {
       padding: 0 2rem;
+
+      .intro {
+        padding: 20px 20px 40px 20px;
+      }
     }
   }
 }
